move CurrentUserProvider to service.security package
because it is mainly concerned with spring security authentication stuff
