+++ src/main/java/org/fairdatateam/fairdatapoint/service/metadata/state/MetadataStateService.java
@@ -31,7 +31,7 @@ import org.fairdatateam.fairdatapoint.entity.metadata.MetadataState;
 import org.fairdatateam.fairdatapoint.entity.resource.ResourceDefinition;
 import org.fairdatateam.fairdatapoint.entity.resource.ResourceDefinitionChild;
 import org.fairdatateam.fairdatapoint.service.metadata.validator.MetadataStateValidator;
-import org.fairdatateam.fairdatapoint.service.user.CurrentUserProvider;
+import org.fairdatateam.fairdatapoint.service.security.CurrentUserProvider;
 import org.eclipse.rdf4j.model.IRI;
 import org.eclipse.rdf4j.model.Model;
 import org.springframework.beans.factory.annotation.Autowired;
